Add error handling and loading indicator to base branch selection
- also puts base branch selection in its own component
